# backend/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Web-App for Managing Student Projects
  version: 1.0.0
...
paths:
...
  /projects:
    get:
      tags: [Project Management, User]
      summary: Retrieve all projects
      description: Admin-only. Returns all projects with summary data.
      operationId: listAllProjects
      responses:
        "200":
          description: List of all projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
        "403":
          description: Only admins are allowed to view all projects.
    post:
      tags: [Project Management, User]
      summary: Create a new project
      description: Admin-only. Creates a project and returns the full project details.
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "400":
          description: Invalid input data.
        "403":
          description: Only admins can create projects.
  /projects/mine:
    get:
      tags: [Project Management, User]
      summary: Retrieve authenticated user's projects
      description: Shows projects where the logged-in user is a manager or developer.
      operationId: listMyProjects
      responses:
        "200":
          description: User's projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
        "403":
          description: The user is not authorized to view these projects.
...
  /projects/{projectId}:
    parameters:
      - in: path
        name: projectId
        schema:
          type: integer
        required: true
        description: Unique ID of the project.
    get:
      tags: [Project Management, User]
      summary: Get project details
      description: Only admins or assigned managers/developers may access.
      operationId: getProjectDetails
      responses:
        "200":
          description: Project details retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "403":
          description: User is not allowed to view this project.
        "404":
          description: Project not found.
    put:
      tags: [Project Management, User]
      summary: Edit an existing project
      description: Only project managers can edit project data.
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "403":
          description: Only project managers can edit.
        "404":
          description: Project not found.
    delete:
      tags: [Project Management, User]
      summary: Delete a project
      description: Admin-only. Deletes the selected project.
      operationId: deleteProject
      responses:
        "200":
          description: Project deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  deletedProject:
                    $ref: "#/components/schemas/ProjectSummary"
        "403":
          description: Only admins can delete projects.
        "404":
          description: Project not found.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api-key:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    ProjectType:
      type: string
      enum:
        - "bachelor’s thesis"
        - "master’s thesis"
        - "advanced internship"
        - "project work"
        - "seminar paper"
        - "independent scientific work"
      description: Type/category of the project.
    ProjectStatus:
      type: string
      enum: [proposed, registered, evaluated, completed]
      description: Current status of the project.
    ProjectSummary:
      type: object
      required: [id, name, type, status, managerEmail]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
        type:
          $ref: "#/components/schemas/ProjectType"
        status:
          $ref: "#/components/schemas/ProjectStatus"
        managerEmail:
          type: string
          format: email
    ProjectDetail:
      type: object
      required: [id, name, type, status, managers, developers]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
        type:
          $ref: "#/components/schemas/ProjectType"
        status:
          $ref: "#/components/schemas/ProjectStatus"
        description:
          type: string
        summary_ai:
          type: string
          description: Overview created by AI according to the last update on the project aggregate, this value is computed every time a change is commited to the project aggregate
        general_sentiment_ai:
          type: string
          description: What do the majority of comments say about the project, this value is computed every time a change is commited to the comment list
        managers:
          type: array
          items:
            type: string
            format: email
        developers:
          type: array
          items:
            type: string
            format: email
    ProjectInput:
      type: object
      required: [name, type, status, managers]
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/ProjectType"
        status:
          $ref: "#/components/schemas/ProjectStatus"
        description:
          type: string
        managers:
          type: array
          items:
            type: string
            format: email
        developers:
          type: array
          items:
            type: string
            format: email
...